contract client moved to test harness dir
