+++ devtools/gammaserver/deploy_native_test.go
@@ -2,6 +2,7 @@ package gammaserver
 
 import (
 	"context"
+	contractClient "github.com/orbs-network/orbs-network-go/test/harness/contracts"
 	"github.com/orbs-network/orbs-network-go/instrumentation/log"
 	"github.com/orbs-network/orbs-network-go/test"
 	"github.com/orbs-network/orbs-network-go/test/contracts"
@@ -17,18 +18,19 @@ func TestNonLeaderDeploysNativeContract(t *testing.T) {
 
 	test.WithContext(func(ctx context.Context) {
 		network := NewDevelopmentNetwork(ctx, log.GetLogger())
+		contract := contractClient.NewContractClient(network)
 
 		counterStart := contracts.MOCK_COUNTER_CONTRACT_START_FROM
 
 		t.Log("deploying contract")
 
-		<-network.GetCounterContract().SendDeployCounterContract(ctx, 1) // leader is nodeIndex 0, validator is nodeIndex 1
-		require.EqualValues(t, counterStart, <-network.GetCounterContract().CallCounterGet(ctx, 0), "get counter after deploy")
+		<-contract.SendDeployCounterContract(ctx, 1) // leader is nodeIndex 0, validator is nodeIndex 1
+		require.EqualValues(t, counterStart, <-contract.CallCounterGet(ctx, 0), "get counter after deploy")
 
 		t.Log("transacting with contract")
 
-		<-network.GetCounterContract().SendCounterAdd(ctx, 1, 17)
-		require.EqualValues(t, counterStart+17, <-network.GetCounterContract().CallCounterGet(ctx, 0), "get counter after transaction")
+		<-contract.SendCounterAdd(ctx, 1, 17)
+		require.EqualValues(t, counterStart+17, <-contract.CallCounterGet(ctx, 0), "get counter after transaction")
 
 	})
 	time.Sleep(5 * time.Millisecond) // give context dependent goroutines 5 ms to terminate gracefully
