+++ test/acceptance/deploy_native_test.go
@@ -2,6 +2,7 @@ package acceptance
 
 import (
 	"context"
+	contractClient "github.com/orbs-network/orbs-network-go/test/harness/contracts"
 	"github.com/orbs-network/orbs-network-go/test/contracts"
 	"github.com/orbs-network/orbs-network-go/test/harness"
 	"github.com/stretchr/testify/require"
@@ -15,16 +16,17 @@ func TestNonLeaderDeploysNativeContract(t *testing.T) {
 
 		counterStart := contracts.MOCK_COUNTER_CONTRACT_START_FROM
 		network.MockContract(contracts.MockForCounter(), string(contracts.NativeSourceCodeForCounter(counterStart)))
+		contract := contractClient.NewContractClient(network)
 
 		t.Log("deploying contract")
 
-		<-network.GetCounterContract().SendDeployCounterContract(ctx, 1)
-		require.EqualValues(t, counterStart, <-network.GetCounterContract().CallCounterGet(ctx, 0), "get counter after deploy")
+		<-contract.SendDeployCounterContract(ctx, 1)
+		require.EqualValues(t, counterStart, <-contract.CallCounterGet(ctx, 0), "get counter after deploy")
 
 		t.Log("transacting with contract")
 
-		<-network.GetCounterContract().SendCounterAdd(ctx, 1, 17)
-		require.EqualValues(t, counterStart+17, <-network.GetCounterContract().CallCounterGet(ctx, 0), "get counter after transaction")
+		<-contract.SendCounterAdd(ctx, 1, 17)
+		require.EqualValues(t, counterStart+17, <-contract.CallCounterGet(ctx, 0), "get counter after transaction")
 
 	})
 }
