+++ test/harness/contracts/api.go
@@ -2,7 +2,6 @@ package contracts
 
 import (
 	"context"
-	"github.com/orbs-network/orbs-network-go/instrumentation/log"
 	"github.com/orbs-network/orbs-spec/types/go/primitives"
 	"github.com/orbs-network/orbs-spec/types/go/protocol"
 	"github.com/orbs-network/orbs-spec/types/go/protocol/client"
@@ -17,9 +16,8 @@ type ContractAPI interface {
 
 type contractClient struct {
 	API    ContractAPI
-	logger log.BasicLogger
 }
 
-func NewContractClient(api ContractAPI, logger log.BasicLogger) *contractClient {
-	return &contractClient{API: api, logger: logger}
+func NewContractClient(api ContractAPI) *contractClient {
+	return &contractClient{API: api}
 }
