+++ test/harness/contracts/benchmark_token.go
