+++ test/harness/contracts/counter.go
